+++ package.json
@@ -86,7 +86,7 @@
     "@types/js-yaml": "^4.0.0",
     "@types/jsonwebtoken": "^8.5.1",
     "@types/lodash": "^4.14.168",
-    "@types/luxon": "^1.26.3",
+    "@types/luxon": "^1.26.4",
     "@types/mjml-react": "^1.0.6",
     "@types/node": "^14.14.37",
     "@types/react": "^16.14.2",
